Major change - claims data
Most elements displayed in the dashboard were impacted by this change. We shifted from evaluations of enrollment data to instead using claims data. Although the claims dataset is not update as regularly as enrollment, the increasing gap between number of enrolled vs. claimed subscribers of ACP indicates that not all households enrolled are actively using the benefit.